requirements.txt
geopandas==0.8.1
pydantic==1.6.1
httpx==0.15.5
